+++ packages/riverpod_cli/lib/src/migrate.dart
@@ -3,7 +3,6 @@ import 'dart:io';
 import 'package:args/command_runner.dart';
 import 'package:codemod/codemod.dart';
 import 'package:glob/glob.dart';
-import 'package:pub_semver/pub_semver.dart';
 import 'package:pubspec_parse/pubspec_parse.dart';
 
 import 'migrate/errors.dart';
@@ -53,19 +52,29 @@ class MigrateCommand extends Command<void> {
       stderr.writeln(
         'Pubspec not found! Are you in the root directory of your package / app?',
       );
-      return;
+      exit(-1);
     }
+
     final pubspec = Pubspec.parse(pubspecFile.readAsStringSync());
     final dep = pubspec.dependencies['hooks_riverpod'] ??
         pubspec.dependencies['flutter_riverpod'] ??
         pubspec.dependencies['riverpod'];
 
-    VersionConstraint version;
-    if (dep is HostedDependency) {
-      version = dep.version;
-    } else {
-      throw UnimplementedError(
-          'Migrating git and path dependencies can cause issues because of trying to understand riverpod versioning, please depend on an official package');
+    if (dep is! HostedDependency) {
+      stderr.writeln(
+        'Migrating git and path dependencies can cause issues because of '
+        'trying to understand riverpod versioning, please depend on an official package',
+      );
+      exit(-1);
+    }
+
+    if (dep.version.allows(latestVersion)) {
+      stderr.writeln(
+        'It seems like your project already has Riverpod $latestVersion installed.\n'
+        'The migration tool will not work if your project already uses the new '
+        'version. To fix, downgrade the version of Riverpod then try again.',
+      );
+      exit(-1);
     }
 
     await runInteractiveCodemodSequence(
@@ -74,11 +83,11 @@ class MigrateCommand extends Command<void> {
         aggregate(
           [
             RiverpodImportAllMigrationSuggestor(),
-            RiverpodNotifierChangesMigrationSuggestor(version),
+            RiverpodNotifierChangesMigrationSuggestor(dep.version),
           ],
         ),
-        RiverpodProviderUsageInfo(version),
-        RiverpodUnifiedSyntaxChangesMigrationSuggestor(version),
+        RiverpodProviderUsageInfo(dep.version),
+        RiverpodUnifiedSyntaxChangesMigrationSuggestor(dep.version),
       ],
       args: argResults?.arguments ?? [],
     );