+++ packages/riverpod_cli/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.0.3
+
+Logged more errors, to help debug some potential issues
+
+# 1.0.2
+
+The command `riverpod migrate` will now fail if used on a project that already
+uses the latest version available.
+
 # 1.0.1
 
 - Updated to migrate to version 1.0.1